Conflicts:
	src/main/java/redis/clients/jedis/JedisCluster.java
	src/test/java/redis/clients/jedis/tests/ShardedJedisTest.java

+++ src/main/java/redis/clients/jedis/BasicCommands.java
@@ -34,7 +34,7 @@ public interface BasicCommands {
 
   String slaveofNoOne();
 
-  Long getDB();
+  int getDB();
 
   String debug(DebugParams params);
